Reapply e642dcde with changes to support fetching schema from remote URL
e642dcde added a helper fixed selection for the item type field.

However, it was not working when the schema was fetched from a remote URL.

This was reported to json-editor: https://github.com/json-editor/json-editor/issues/1165

This commit is a workaround which does not use $ref, hence duplicating some code.