Revert "Remove no longer necessary @types/jsdom package"
This reverts commit 7d8b7f442e22aca6fc6b3e080ee29db088f3f741.

I did not see jsdom as a dependency and thought that we had left this
types dependency from when we used jsdom in w2c-core.

However, jsdom is still a development dependency because we use it
in automatic tests (we need a way to simulate the window object
that is passed to the Domain object constructor).